requirements.txt
flask
requests
gunicorn
flask-cors
authlib
python-dotenv